+++ Sources/Instructions/Core/Public/CoachMarksController.swift
@@ -1,9 +1,9 @@
 // CoachMarksController.swift
 //
-// Copyright (c) 2015, 2016 Frédéric Maquin <fred@ephread.com>,
-//                          Daniel Basedow <daniel.basedow@gmail.com>,
-//                          Esteban Soto <esteban.soto.dev@gmail.com>,
-//                          Ogan Topkaya <>
+// Copyright (c) 2015-2016, 2018 Frédéric Maquin <fred@ephread.com>,
+//                               Daniel Basedow <daniel.basedow@gmail.com>,
+//                               Esteban Soto <esteban.soto.dev@gmail.com>,
+//                               Ogan Topkaya <>
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -37,10 +37,41 @@ public class CoachMarksController {
     /// be called at various points.
     public weak var delegate: CoachMarksControllerDelegate?
 
+    /// Implement the animation delegate protocol, which methods will
+    /// be called at various points.
+    public weak var animationDelegate: CoachMarksControllerAnimationDelegate?
+
+    /// Controls the style of the status bar when coach marks are displayed
+    public var statusBarStyle: UIStatusBarStyle? {
+        get {
+            return coachMarksViewController.customStatusBarStyle
+        }
+
+        set {
+            coachMarksViewController.customStatusBarStyle = newValue
+            coachMarksViewController.setNeedsStatusBarAppearanceUpdate()
+        }
+    }
+
+    public var rotationStyle: RotationStyle {
+        get { return coachMarksViewController.rotationStyle }
+        set { coachMarksViewController.rotationStyle = newValue }
+    }
+
+    public var statusBarVisibility: StatusBarVisibility {
+        get { return coachMarksViewController.statusBarVisibility }
+        set { coachMarksViewController.statusBarVisibility = newValue }
+    }
+
+    public var interfaceOrientations: InterfaceOrientations {
+        get { return coachMarksViewController.interfaceOrientations }
+        set { coachMarksViewController.interfaceOrientations = newValue }
+    }
+
     /// Hide the UI.
     fileprivate(set) public lazy var overlay: OverlayManager = {
         let overlay = OverlayManager()
-        overlay.delegate = self
+        overlay.overlayDelegate = self
 
         return overlay
     }()
@@ -65,14 +96,14 @@ public class CoachMarksController {
 
     // MARK: - Private properties
     fileprivate weak var controllerWindow: UIWindow?
-
     fileprivate var coachMarksWindow: UIWindow?
 
     /// Handle the UI part
     fileprivate lazy var coachMarksViewController: CoachMarksViewController = {
-        let coachMarkController = CoachMarksViewController()
-        coachMarkController.coachMarkDisplayManager = self.buildCoachMarkDisplayManager()
-        coachMarkController.skipViewDisplayManager = self.buildSkipViewDisplayManager()
+        let coachMarkController = CoachMarksViewController(
+            coachMarkDisplayManager: self.buildCoachMarkDisplayManager(),
+            skipViewDisplayManager: self.buildSkipViewDisplayManager()
+        )
 
         coachMarkController.overlayManager = self.overlay
 
@@ -94,70 +125,84 @@ public extension CoachMarksController {
 
 // MARK: - Flow management
 public extension CoachMarksController {
-    /// Start displaying the coach marks.
+    /// Start instructions in the given context.
     ///
-    /// - Parameter parentViewController: View Controller to which attach self.
-    public func start(on parentViewController: UIViewController) {
+    /// - Parameter presentationContext: the context in which show Instructions
+
+    func start(in presentationContext: PresentationContext) {
         guard let dataSource = self.dataSource else {
             print("startOn: snap! you didn't setup any datasource, the" +
-                  "coach mark manager won't do anything.")
+                "coach mark manager won't do anything.")
             return
         }
 
-        // If coach marks are currently being displayed, calling `start()` doesn't do anything.
+        // If coach marks are currently being displayed, calling `start(in: )` doesn't do anything.
         if flow.started { return }
 
         let numberOfCoachMarks = dataSource.numberOfCoachMarks(for: self)
         if numberOfCoachMarks <= 0 {
             print("startOn: the dataSource returned an invalid value for " +
-                  "numberOfCoachMarksForCoachMarksController(_:)")
+                "numberOfCoachMarksForCoachMarksController(_:)")
             return
         }
 
+        switch presentationContext {
+        case .newWindow(let viewController, let windowLevel):
 #if INSTRUCTIONS_APP_EXTENSIONS
-        coachMarksViewController.attach(to: parentViewController)
+            fatalError("PresentationContext.newWindow(above:) is not available in App Extensions.")
 #else
-        controllerWindow = parentViewController.view.window
-        coachMarksWindow = coachMarksWindow ?? InstructionsWindow(frame: UIScreen.main.bounds)
-
-        coachMarksViewController.attach(to: coachMarksWindow!, of: parentViewController)
+            controllerWindow = viewController.view.window
+            coachMarksWindow = coachMarksWindow ?? InstructionsWindow(frame: UIScreen.main.bounds)
+            coachMarksViewController.attach(to: coachMarksWindow!, over: viewController,
+                                            at: windowLevel)
 #endif
+        case .currentWindow(let viewController):
+            coachMarksViewController.attachToWindow(of: viewController)
+        case .viewController(let viewController):
+            coachMarksViewController.attach(to: viewController)
+        }
+
+        delegate?.coachMarksController(self,
+                                       configureOrnamentsOfOverlay: overlay.overlayView.ornaments)
+
         flow.startFlow(withNumberOfCoachMarks: numberOfCoachMarks)
+    }
+
+    /// Start displaying the coach marks.
+    ///
+    /// - Parameter parentViewController: View Controller to which attach self.
+    @available(iOS, deprecated: 1.2.1, message: "use start(in:) instead.")
+    func start(on parentViewController: UIViewController) {
+#if INSTRUCTIONS_APP_EXTENSIONS
+        start(in: .currentWindow(of: parentViewController))
+#else
+        start(in: .newWindow(over: parentViewController, at: nil))
+#endif
     }
 
     /// Stop the flow of coach marks. Don't forget to call this method in viewDidDisappear or
     /// viewWillDisappear.
     ///
     /// - Parameter immediately: `true` to stop immediately, without animations.
-    public func stop(immediately: Bool = false) {
+    func stop(immediately: Bool = false) {
         if immediately {
-            flow.stopFlow(immediately: true, userDidSkip: false, shouldCallDelegate: false)
+            flow.stopFlow(immediately: true, userDidSkip: false,
+                          shouldCallDelegate: false) { [weak self] in
+                self?.coachMarksWindow = nil
+            }
         } else {
-            flow.stopFlow()
+            flow.stopFlow { [weak self] in
+                self?.coachMarksWindow = nil
+            }
         }
     }
 
-    public func prepareForChange() {
+    func prepareForChange() {
         coachMarksViewController.prepareForChange()
     }
 
-    public func restoreAfterChangeDidComplete() {
+    func restoreAfterChangeDidComplete() {
         coachMarksViewController.restoreAfterChangeDidComplete()
-    }
-
-    /// Pause the display.
-    /// This method is expected to be used by the delegate to
-    /// stop the display, perform animation and resume display with `resume()`
-    @available(*, deprecated: 0.6.0, message: "Please use flow.pause() instead.")
-    func pause() {
-        flow.pause()
-    }
-
-    /// Resume the display.
-    /// If the display wasn't paused earlier, this method won't do anything.
-    @available(*, deprecated: 0.6.0, message: "Please use flow.resume() instead.")
-    func resume() {
-        flow.resume()
     }
 }
 
@@ -183,6 +228,7 @@ private extension CoachMarksController {
         let coachMarkDisplayManager =
             CoachMarkDisplayManager(coachMarkLayoutHelper: CoachMarkLayoutHelper())
         coachMarkDisplayManager.dataSource = self
+        coachMarkDisplayManager.animationDelegate = self
 
         return coachMarkDisplayManager
     }