+++ Sources/Instructions/Managers/Public/FlowManager.swift
@@ -1,6 +1,6 @@
 // FlowManager.swift
 //
-// Copyright (c) 2016 Frédéric Maquin <fred@ephread.com>
+// Copyright (c) 2016, 2018 Frédéric Maquin <fred@ephread.com>
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -77,6 +77,7 @@ public class FlowManager {
         self.coachMarksViewController = coachMarksViewController
     }
 
+    // MARK: Internal methods
     internal func startFlow(withNumberOfCoachMarks numberOfCoachMarks: Int) {
         disableFlow = false
 
@@ -85,17 +86,6 @@ public class FlowManager {
         coachMarksViewController.prepareToShowCoachMarks {
             self.showNextCoachMark()
         }
-    }
-
-    public func resume() {
-        if started && paused {
-            paused = false
-            createAndShowCoachMark(afterResuming: true)
-        }
-    }
-
-    public func pause() {
-        paused = true
     }
 
     internal func reset() {
@@ -113,7 +103,7 @@ public class FlowManager {
     /// - Parameter shouldCallDelegate: `true` to notify the delegate that the flow
     ///                                 was stop.
     internal func stopFlow(immediately: Bool = false, userDidSkip skipped: Bool = false,
-                           shouldCallDelegate: Bool = true ) {
+                           shouldCallDelegate: Bool = true, completion: (() -> Void)? = nil) {
         reset()
 
         let animationBlock = { () -> Void in
@@ -121,9 +111,11 @@ public class FlowManager {
             self.coachMarksViewController.currentCoachMarkView?.alpha = 0.0
         }
 
-        let completionBlock = {(finished: Bool) -> Void in
-            self.coachMarksViewController.detachFromWindow()
-            if shouldCallDelegate { self.delegate?.didEndShowingBySkipping(skipped) }
+        let completionBlock = { [weak self] (finished: Bool) -> Void in
+            guard let strongSelf = self else { return }
+            strongSelf.coachMarksViewController.detachFromWindow()
+            if shouldCallDelegate { strongSelf.delegate?.didEndShowingBySkipping(skipped) }
+            completion?()
         }
 
         if immediately {
@@ -133,15 +125,19 @@ public class FlowManager {
             // TODO: SoC
             self.coachMarksViewController.overlayManager.overlayView.alpha = 0
         } else {
-            UIView.animate(withDuration: coachMarksViewController.overlayManager.fadeAnimationDuration,
-                                         animations: animationBlock)
+            UIView.animate(withDuration: coachMarksViewController.overlayManager
+                                                                 .fadeAnimationDuration,
+                           animations: animationBlock)
 
-            self.coachMarksViewController.overlayManager.showOverlay(false, completion: completionBlock)
+            self.coachMarksViewController.overlayManager.showOverlay(false,
+                                                                     completion: completionBlock)
         }
     }
 
     internal func showNextCoachMark(hidePrevious: Bool = true) {
         if disableFlow || paused || !canShowCoachMark { return }
+
+        let previousIndex = currentIndex
 
         canShowCoachMark = false
         currentIndex += 1
@@ -158,7 +154,36 @@ public class FlowManager {
         if hidePrevious {
             guard let currentCoachMark = currentCoachMark else { return }
 
-            coachMarksViewController.hide(coachMark: currentCoachMark) {
+            coachMarksViewController.hide(coachMark: currentCoachMark, at: previousIndex) {
+                self.delegate?.didHide(coachMark: self.currentCoachMark!, at: self.currentIndex - 1)
+                self.showOrStop()
+            }
+        } else {
+            showOrStop()
+        }
+    }
+
+    internal func showPreviousCoachMark(hidePrevious: Bool = true) {
+        if disableFlow || paused || !canShowCoachMark { return }
+
+        let previousIndex = currentIndex
+
+        canShowCoachMark = false
+        currentIndex -= 1
+
+        if currentIndex < 0 {
+            stopFlow()
+            return
+        }
+
+        if let currentCoachMark = currentCoachMark {
+            delegate?.willHide(coachMark: currentCoachMark, at: currentIndex + 1)
+        }
+
+        if hidePrevious {
+            guard let currentCoachMark = currentCoachMark else { return }
+
+            coachMarksViewController.hide(coachMark: currentCoachMark, at: previousIndex) {
                 self.delegate?.didHide(coachMark: self.currentCoachMark!, at: self.currentIndex)
                 self.showOrStop()
             }
@@ -181,7 +206,7 @@ public class FlowManager {
     ///
     /// - Parameter shouldCallDelegate: `true` to call delegate methods, `false` otherwise.
     internal func createAndShowCoachMark(afterResuming: Bool = false,
-                                         recreatedAfterTransition recreated: Bool = false) {
+                                         changing change: ConfigurationChange = .nothing) {
         if disableFlow { return }
 
         if !afterResuming {
@@ -198,7 +223,11 @@ public class FlowManager {
             // The coach mark will soon show, we notify the delegate, so it
             // can perform some things and, if required, update the coach mark structure.
             self.delegate?.willShow(coachMark: &currentCoachMark!,
-                                    afterSizeTransition: recreated, at: currentIndex)
+                                    beforeChanging: change, at: currentIndex)
+
+            // TODO: ❗️ To remove in 2.0.0
+            self.delegate?.willShow(coachMark: &currentCoachMark!,
+                                    afterSizeTransition: (change == .size), at: currentIndex)
         }
 
         // The delegate might have paused the flow, he check whether or not it's
@@ -215,8 +244,44 @@ public class FlowManager {
                 self.canShowCoachMark = true
 
                 self.delegate?.didShow(coachMark: self.currentCoachMark!,
-                                       afterSizeTransition: recreated, at: self.currentIndex)
+                                       afterChanging: change, at: self.currentIndex)
+
+                // TODO: ❗️ To remove in 2.0.0
+                self.delegate?.didShow(coachMark: self.currentCoachMark!,
+                                       afterSizeTransition: (change == .size),
+                                       at: self.currentIndex)
             }
+        }
+    }
+
+    // MARK: Public methods
+    public func resume() {
+        if started && paused {
+            paused = false
+
+            let completion: (Bool) -> Void = { _ in
+                self.createAndShowCoachMark(afterResuming: true)
+            }
+
+            if coachMarksViewController.overlayManager.isWindowHidden {
+                coachMarksViewController.overlayManager.showWindow(true, completion: completion)
+            } else if coachMarksViewController.overlayManager.isOverlayInvisible {
+                coachMarksViewController.overlayManager.showOverlay(true, completion: completion)
+            } else {
+                completion(true)
+            }
+        }
+    }
+
+    public func pause(and pauseStyle: PauseStyle = .hideNothing) {
+        paused = true
+
+        switch pauseStyle {
+        case .hideInstructions:
+            coachMarksViewController.overlayManager.showWindow(false, completion: nil)
+        case .hideOverlay:
+            coachMarksViewController.overlayManager.showOverlay(false, completion: nil)
+        case .hideNothing: break
         }
     }
 
@@ -224,7 +289,7 @@ public class FlowManager {
     ///
     /// - Parameter numberOfCoachMarksToSkip: the number of coach marks
     ///                                       to skip.
-    open func showNext(numberOfCoachMarksToSkip numberToSkip: Int = 0) {
+    public func showNext(numberOfCoachMarksToSkip numberToSkip: Int = 0) {
         if !self.started || !canShowCoachMark { return }
 
         if numberToSkip < 0 {
@@ -236,6 +301,24 @@ public class FlowManager {
         currentIndex += numberToSkip
 
         showNextCoachMark(hidePrevious: true)
+    }
+
+    /// Show the previous specified Coach Mark.
+    ///
+    /// - Parameter numberOfCoachMarksToSkip: the number of coach marks
+    ///                                       to skip.
+    public func showPrevious(numberOfCoachMarksToSkip numberToSkip: Int = 0) {
+        if !self.started || !canShowCoachMark { return }
+
+        if numberToSkip < 0 {
+            print("showPrevious: The specified number of coach marks to skip" +
+                  "was negative, nothing to do.")
+            return
+        }
+
+        currentIndex -= numberToSkip
+
+        showPreviousCoachMark(hidePrevious: true)
     }
 }
 
@@ -251,13 +334,13 @@ extension FlowManager: CoachMarksViewControllerDelegate {
     func willTransition() {
         coachMarksViewController.prepareForSizeTransition()
         if let coachMark = currentCoachMark {
-            coachMarksViewController.hide(coachMark: coachMark, animated: false,
-                                          beforeTransition: true)
+            coachMarksViewController.hide(coachMark: coachMark, at: currentIndex,
+                                          animated: false, beforeTransition: true)
         }
     }
 
-    func didTransition() {
+    func didTransition(afterChanging change: ConfigurationChange) {
         coachMarksViewController.restoreAfterSizeTransitionDidComplete()
-        createAndShowCoachMark(recreatedAfterTransition: true)
+        createAndShowCoachMark(changing: change)
     }
 }